python-agents/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.15.0
python-dotenv>=0.19.0
crewai>=0.95.0
langchain>=0.1.0
langchain-openai>=0.0.2
typing-extensions>=4.5.0
pydantic>=2.4.2
openai>=1.1.0 